requirements.txt
pyquery
requests
awesome-slugify
